% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11sampling-package.R
\name{bootstrap_variable_cpp11}
\alias{bootstrap_variable_cpp11}
\title{Bootstrap Resampling with Variable Sample Sizes (cpp11 version)}
\usage{
bootstrap_variable_cpp11(data, min_size, max_size, n_bootstrap)
}
\arguments{
\item{data}{Original data vector to bootstrap from}

\item{min_size}{Minimum bootstrap sample size}

\item{max_size}{Maximum bootstrap sample size}

\item{n_bootstrap}{Number of bootstrap samples to generate}
}
\value{
A list of bootstrap samples, each with a different random size
}
\description{
Performs bootstrap resampling where each bootstrap sample has a randomly
  determined size. Demonstrates efficient dynamic vector growth with cpp11.
}
