% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11sampling-package.R
\name{rejection_sampling_cpp11}
\alias{rejection_sampling_cpp11}
\title{Rejection Sampling from Truncated Normal Distribution (cpp11 version)}
\usage{
rejection_sampling_cpp11(n_samples, mu = 0, sigma = 1, lower = -2, upper = 2)
}
\arguments{
\item{n_samples}{Number of accepted samples desired}

\item{mu}{Mean of the normal distribution. Default is 0.}

\item{sigma}{Standard deviation of the normal distribution. Default is 1.}

\item{lower}{Lower truncation bound. Default is -2.}

\item{upper}{Upper truncation bound. Default is 2.}
}
\value{
A numeric vector of accepted samples from the truncated normal distribution
}
\description{
Generates samples from a truncated normal distribution using rejection sampling.
  This cpp11 implementation demonstrates efficient vector growth with push_back operations.
}
